pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyazsentinel"
version = "0.1.0"
description = "Python SDK and CLI for Azure Sentinel management"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "misterapol", email = "dev@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]
keywords = ["azure", "sentinel", "security", "siem"]
requires-python = ">=3.8"
dependencies = [
    "azure-identity>=1.12.0",
    "azure-mgmt-resource>=21.0.0",
    "azure-mgmt-loganalytics>=12.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "PyYAML>=6.0",
    "requests>=2.28.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "flake8>=5.0.0",
    "pre-commit>=2.20.0",
]

[project.urls]
Homepage = "https://github.com/misterapol/PyAZSentinel"
Repository = "https://github.com/misterapol/PyAZSentinel"
Issues = "https://github.com/misterapol/PyAZSentinel/issues"

[project.scripts]
pyazsentinel = "pyazsentinel.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pyazsentinel*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=pyazsentinel --cov-report=term-missing"
